requirements.txt
pandas
numpy
openpyxl
Flask
fastapi[standard]
langchain
langgraph
langsmith
langchain-core
langchain-google-genai
google-generativeai
notebook
ipykernel
pillow
streamlit
langgraph-cli[inmem]
langgraph-api
streamlit_chat